Avoid externalizing new internal implementation classes as API.
