+++ api/src/main/java/jakarta/data/metamodel/TextAttribute.java
@@ -19,6 +19,7 @@ package jakarta.data.metamodel;
 
 import jakarta.data.Sort;
 import jakarta.data.metamodel.constraint.Like;
+import jakarta.data.metamodel.impl.TextAttributeRecord;
 import jakarta.data.metamodel.restrict.Restrict;
 import jakarta.data.metamodel.restrict.TextRestriction;
 
@@ -95,6 +96,21 @@ public interface TextAttribute<T> extends ComparableAttribute<T,String> {
 
     default TextRestriction<T> notEndsWith(String suffix) {
         return Restrict.notEndsWith(suffix, name());
+    }
+
+    /**
+     * <p>Creates a static metamodel {@code TextAttribute} representing the
+     * entity attribute with the specified name.</p>
+     *
+     * @param <T> entity class of the static metamodel.
+     * @param name the name of the entity attribute.
+     * @return instance of {@code TextAttribute}.
+     */
+    static <T> TextAttribute<T> of(String name) {
+        if (name == null)
+            throw new IllegalArgumentException("entity attribute name is required");
+
+        return new TextAttributeRecord<>(name);
     }
 
     default TextRestriction<T> startsWith(String prefix) {
